Issue #10 - added component and files for second level domain under .tld part 3

+++ entrypoints/entrypoint_knotd.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-if [ "$HOSTNAME" != knot-fakeroot ]; then
+if [ "$HOSTNAME" != knot-fakeroot ] && [ "$HOSTNAME" != knot-secondlevel ]; then
    rm /var/lib/knot/zones/* /var/lib/knot/journal/* > /dev/null 2>&1
 fi
 chown --recursive knot:knot /run/knot/ /var/lib/knot /etc/knot
